fix(PoolSuggestControl): load all pools
